Refactor member service to use UUIDs for member identification
- Updated MemberServiceClient and MemberServiceServer interfaces to remove comments about RegisterMember.
- Added UUID dependency in go.mod.
- Modified member converters to handle UUIDs instead of int64.
- Updated handlers to convert member IDs to and from UUIDs.
- Refactored ListMembers method to use UUID cursor for pagination.
- Changed SQL queries to use UUIDs for member identification.
- Updated service methods to accept and return UUIDs for member operations.
- Created utility functions for UUID conversion.
- Adjusted SQL migrations to create members and profiles with UUIDs.
- Updated protobuf definitions to reflect changes in member ID types.

+++ migrations/20260706072001_seed_default_branch.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 INSERT INTO member_branch_counters (branch_id, next_member_number)
-VALUES (1, 1)
+VALUES (1::BIGINT, 1::BIGINT)
 ON CONFLICT (branch_id) DO NOTHING;
 
 -- +goose Down
